Refactor direcctorios, eliminado Panel
