+++ GestorProject/src/App/Main.java
@@ -1,13 +1,7 @@
 package App;
 
-import Gestores.GestoreGUI.*;
-
-import javax.swing.*;
-
 public class Main {
     public static void main(String args[]){
         System.out.println("Hello world");
-
     }
-
 }
