--- GestorProject/src/GUI/Panel.java
@@ -1,15 +1,0 @@
-package GUI;
-
-import javax.swing.*;
-
-public class Panel extends JPanel {
-    protected String titulo;
-
-    public Panel(String titulo) {
-        this.titulo = titulo;
-    }
-
-    public String getTitulo() {
-        return titulo;
-    }
-}
