+++ GestorProject/src/GUI/PanelWrap.java
@@ -1,4 +1,4 @@
-package Gestores.GestoreGUI;
+package GUI;
 
 import java.util.ArrayList;
 
@@ -10,7 +10,7 @@ public class PanelWrap {
         this.nombre = nombre;
         this.clase = clase;
     }
-    public static PanelWrap find (String nombre, ArrayList<PanelWrap> paneles) {
+    public static PanelWrap find(String nombre, ArrayList<PanelWrap> paneles) {
         for (PanelWrap panel : paneles) {
             if (panel.nombre.equals(nombre)) return panel;
         }
