+++ GestorProject/src/TestGUI/panelTest1.java
@@ -1,20 +1,14 @@
-package Gestores.GestoreGUI.GestorGUITest;
+package TestGUI;
 
-import GUI.Panel;
-import Gestores.GestoreGUI.GestorGUI;
+import GUI.GestorGUI;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class panelTest1 extends Panel {
+public class panelTest1 extends JPanel {
 
-    public panelTest1(String titulo) {
-        super(titulo);
-
-    }
     public panelTest1(){
-        this("Panel 1");
 
         JButton panel2 = new JButton("ir panel 2");
         JButton panel3 = new JButton("ir panel 3");
@@ -31,6 +25,8 @@ public class panelTest1 extends Panel {
                 GestorGUI.push("panel3");
             }
         });
-        add(panel2);add(panel3);
+        add(panel2);
+        add(panel3);
+
     }
 }
